Added support for bundler (Gemfile)
Makes it easier for others to contribute without having to hunt for
dependencies to get the tests to pass. Ideally, this Gemfile would be
just a `gemspec` declaration, but that would require
`add_development_dependency` declarations in the gemspec. Unfortunately,
the gemspec generation is heavily coupled to the Rakefile and the
library itself, which makes the addition of development dependencies
worthy of a separate pull-request in itself.